refactor: Firebase 모델 파일 분리
