+++ Letports/Letports/Common/GatheringMember.swift
@@ -1,0 +1,27 @@
+//
+//  GatheringMebmer.swift
+//  Letports
+//
+//  Created by mosi on 8/22/24.
+//
+import Foundation
+
+struct GatheringMember: Codable {
+    let answer: String
+    let image: String
+    let joinDate: String
+    let joinStatus: String
+    let nickName: String
+    let userUID: String
+    let simpleInfo: String
+    
+    enum CodingKeys: String, CodingKey {
+        case answer = "Answer"
+        case image = "Image"
+        case joinDate = "JoinDate"
+        case joinStatus = "JoinStatus"
+        case nickName = "NickName"
+        case userUID = "UserUID"
+        case simpleInfo = "SimpleInfo"
+    }
+}
