+++ Letports/Letports/Common/Sports.swift
@@ -1,0 +1,12 @@
+//
+//  Sports.swift
+//  Letports
+//
+//  Created by mosi on 8/22/24.
+//
+import Foundation
+
+struct Sports: Codable {
+    let sportsUID: String
+    let sportsName: String
+}
